Added Frequently used word/phrase documents + document generator
this is mostly for building framework for a ML program that will process webpage data and find the most optimal data about the ingredient.

otherwise, for now it is simply used to speed up the process of finding the html elements by hand. once we have a comprehensive list, we will search for them and find the most common HTML elements that correlate to that key word

- word